% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search_cv.R
\name{eval_params}
\alias{eval_params}
\title{Evaluate parameters}
\usage{
eval_params(settings, test_mat, pcp_model, test_mask)
}
\arguments{
\item{settings}{Documented parameter settings associated with the given
\code{pcp_model}.}

\item{test_mat}{The given test matrix with \code{test_mask} \emph{observed} (i.e. not
missing).}

\item{pcp_model}{The output of a PCP algorithm, either \code{\link[=rrmc]{rrmc()}} or
\code{\link[=root_pcp]{root_pcp()}}.}

\item{test_mask}{The binary mask indicating which entries in \code{test_mat}
comprise the test set.}
}
\value{
A \code{data.frame} object with PCP's performance statistics.
}
\description{
Given documented parameter settings \code{settings}, a test matrix \code{test_mat},
PCP solution \code{pcp_model}, and binary mask \code{test_mask}, returns statistics
about the \code{pcp_model}'s performance with \code{settings} on the \code{test_mat}.
\strong{This is an internal function needed by \code{\link[=grid_search_cv]{grid_search_cv()}}. It is not
expected that users should require access to this function.}
}
\seealso{
\code{\link[=grid_search_cv]{grid_search_cv()}}
}
\keyword{internal}
