% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pcp_defaults}
\alias{get_pcp_defaults}
\title{Retrieves the default Root PCP parameter settings for a given matrix.}
\usage{
get_pcp_defaults(D)
}
\arguments{
\item{D}{The matrix for which the default lambda and mu are needed.}
}
\value{
A list containing the default lambda and mu values used in most PCP functions, and the default eta value used in \code{\link{RRMC}}. labelled as "lambda", "mu", and "eta" respectively.
}
\description{
\code{get_pcp_defaults} retrieves the default Root PCP lambda and mu parameter settings for a given matrix.
}
\examples{

# simulate a data matrix:

n <- 50
p <- 10
data <- sim_data(sim_seed = 1, nrow = n, ncol = p, rank = 3, sigma=0, add_sparse = FALSE)
D <- data$M

# get the default PCP parameters:
default <- get_pcp_defaults(D)
default_lambda <- default$lambda
default_mu <- default$mu
}
