% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRMC.R
\name{hard_thresholding}
\alias{hard_thresholding}
\title{Hard-thresholding}
\usage{
hard_thresholding(S, c)
}
\arguments{
\item{S}{The input sparse matrix.}

\item{c}{The amount (or threshold) by which hard-thresholding penalizes \code{S}.}
}
\value{
The hard-thresholded sparse matrix.
}
\description{
\code{hard_thresholding} implements hard-thresholding for the sparse matrix.
}
\seealso{
\code{\link{prox_l1}}
}
\keyword{internal}
