% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root_pcp.R
\name{loss_lod}
\alias{loss_lod}
\title{Loss function for PCP's limit of detection penalty}
\usage{
loss_lod(D, X, LOD)
}
\arguments{
\item{D}{The original data matrix.}

\item{X}{The predicted value of \code{L + S} at the current iteration.}

\item{LOD}{The LOD as a matrix, where \code{dim(LOD) == dim(D)}.}
}
\value{
Scalar value used to calculate loss in objective function.
}
\description{
\code{loss_lod()} includes the LOD-specific penalty terms to compute the loss on
the squared error term \code{L + S - D}.
}
\seealso{
\code{\link[=root_pcp]{root_pcp()}}
}
\keyword{internal}
