% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search_cv.R
\name{matrix_rank}
\alias{matrix_rank}
\title{Estimate rank of a given matrix}
\usage{
matrix_rank(D, thresh = NULL)
}
\arguments{
\item{D}{The input data matrix.}

\item{thresh}{(Optional) A double \verb{eqn\{> 0\}}, specifying the relative
threshold by which "practically zero" is determined, used to calculate the
rank of \code{D}. By default, \code{thresh = NULL}, in which case the threshold is
set to \code{max(dim(D)) * .Machine$double.eps}.}
}
\value{
An integer representing the rank of \code{D}.
}
\description{
\code{matrix_rank()} estimates the rank of a given data matrix \code{D} by counting the
number of "practically nonzero" singular values of \code{D}.

A singular value \eqn{s} is determined to be "practically nonzero" if
\eqn{s \geq s_max * thresh}, i.e. if it is greater than or equal to the
maximum singular value in \code{D} scaled by a given threshold \code{thresh}.
}
\examples{
data <- sim_data()
matrix_rank(data$D)
matrix_rank(data$L)
}
\seealso{
\code{\link[=sparsity]{sparsity()}}
}
