% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcpr-package.R
\docType{package}
\name{pcpr-package}
\alias{pcpr}
\alias{pcpr-package}
\title{pcpr: Principal Component Pursuit for Environmental Epidemiology}
\description{
Implementation of the pattern recognition technique Principal Component Pursuit tailored to environmental health data, as described in Gibson et al. (2022) \doi{10.1289/EHP10479}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Columbia-PRIME/pcpr}
  \item \url{https://columbia-prime.github.io/pcpr/}
  \item Report bugs at \url{https://github.com/Columbia-PRIME/pcpr/issues}
}

}
\author{
\strong{Maintainer}: Lawrence G. Chillrud \email{chili@u.northwestern.edu}

Authors:
\itemize{
  \item Jaime Benavides
  \item Elizabeth A. Gibson
  \item Junhui Zhang
  \item Jingkai Yan
  \item John N. Wright
  \item Jeff Goldsmith
  \item Marianthi-Anna Kioumourtzoglou
}

}
\keyword{internal}
