% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRMC.R
\name{proj_rank_r}
\alias{proj_rank_r}
\title{Project rank}
\usage{
proj_rank_r(L, r)
}
\arguments{
\item{L}{The input low-rank matrix.}

\item{r}{The rank that \code{L} should be projected onto / truncated to.}
}
\value{
The best rank r approximation to the input matrix.
}
\description{
\code{proj_rank_r} implements a best (ie closest) rank-r approximation of an input matrix.
This provides the non-convex replacement for nuclear norm on the \code{L} matrix, and is computed
by retaining the first r leading singular values/vectors of \code{L}. This is equivalent to
solving the following optimization problem: min || X - L ||_F s.t. rank(X) <= r.
}
\seealso{
\code{\link{prox_nuclear}}
}
\keyword{internal}
