% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prox_fro}
\alias{prox_fro}
\title{Prox Frobenius}
\usage{
prox_fro(X, c)
}
\arguments{
\item{X}{The error matrix, \code{D-L-S}}

\item{c}{The amount by which the prox Frobenius method penalizes \code{X}.}
}
\value{
The thresholded error matrix.
}
\description{
\code{prox_fro} implements the proximal gradient method for the Frobenius norm.
This thresholding minimizes the square root of the sum of squared error.
The proximal gradient method is used to solve non-differentiable convex optimization problems.
This is only used in \code{rootPCP} because the squared Frobenius error in \code{stablePCP} is differentiable.
}
