% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox_operators.R
\name{prox_frobenius}
\alias{prox_frobenius}
\title{Proximal gradient method for the Frobenius norm}
\usage{
prox_frobenius(Z, c)
}
\arguments{
\item{Z}{The input error/noise matrix, \code{Z = D - L - S}.}

\item{c}{The amount by which the prox Frobenius method penalizes \code{Z}.}
}
\value{
The thresholded error/noise matrix.
}
\description{
\code{prox_frobenius()} implements the proximal gradient method for the Frobenius
norm. The proximal gradient method is used to solve non-differentiable convex
optimization problems. In \code{\link[=root_pcp]{root_pcp()}}, this thresholding minimizes the
square root of the sum of squared error, where the error is defined as
\code{Z = D - L - S}. \strong{This is an internal function needed by \code{\link[=root_pcp]{root_pcp()}}. It is
not expected that users should require access to this function.}
}
\seealso{
\code{\link[=prox_l1]{prox_l1()}}, \code{\link[=prox_nuclear]{prox_nuclear()}}
}
\keyword{internal}
