% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox_operators.R
\name{prox_l1}
\alias{prox_l1}
\title{Prox L1}
\usage{
prox_l1(S, c)
}
\arguments{
\item{S}{The input sparse matrix.}

\item{c}{The amount by which the prox L1 method penalizes \code{S}.}
}
\value{
The thresholded sparse matrix.
}
\description{
\code{prox_l1} implements the proximal gradient method for the L1 norm.
This soft thresholding encourages the \code{S} matrix to be sparse.
The proximal gradient method is used to solve non-differentiable convex optimization problems.
This is used in \code{stable_pcp} and \code{root_pcp}.
}
\seealso{
\code{\link{hard_thresholding}}, \code{\link{prox_frobenius}}, \code{\link{prox_nuclear}}
}
\keyword{internal}
