% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox_operators.R
\name{prox_nuclear}
\alias{prox_nuclear}
\title{Prox Nuclear}
\usage{
prox_nuclear(L, c)
}
\arguments{
\item{L}{The input low-rank matrix.}

\item{c}{The amount by which the prox nuclear method penalizes \code{L}.}
}
\value{
\code{prox_nuclear} returns a list containing the following two objects:
\describe{
   \item{\code{X}}{The thresholded low-rank matrix.}
   \item{\code{X_nuclear_norm}}{The sum of the absolute values of the thresholded singular values (used in objective function).}
}
}
\description{
\code{prox_nuclear} implements the proximal gradient method for the nuclear norm.
The nuclear norm is equivalent to the L2 norm of the singular values of the matrix.
This singular value thresholding encourages the low-rank \code{L} matrix to be low rank.
The proximal gradient method is used to solve non-differentiable convex optimization problems.
This is used in \code{stable_pcp} and \code{root_pcp}.
}
\seealso{
\code{\link{proj_rank_r}}, \code{\link{prox_frobenius}}, \code{\link{prox_l1}}
}
\keyword{internal}
