% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root_pcp.R
\name{root_pcp}
\alias{root_pcp}
\title{Needs a code check!}
\usage{
root_pcp(
  D,
  lambda,
  mu,
  LOD = -Inf,
  non_negative = T,
  max_iter = 10000,
  verbose = F
)
}
\description{
Needs a code check!
}
\keyword{internal}
