% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root_pcp_noncvx_w_na.R
\name{root_pcp_noncvx_na}
\alias{root_pcp_noncvx_na}
\title{Nonconvex squareroot PCP function with missing values (NA)}
\usage{
root_pcp_noncvx_na(D, lambda, mu, r, verbose = FALSE)
}
\arguments{
\item{D}{The original dataset.}

\item{lambda}{The \code{lambda} parameter penalizes the proximal L1 gradient on the \code{S} matrix.}

\item{mu}{The \code{mu} parameter penalizes the error term.}

\item{r}{The \code{r} parameter specifies the desired rank.}

\item{verbose}{A logical indicating if you would like information on the number of iterations required to reach convergence printed. Optional, and by default \code{verbose = FALSE}.}
}
\value{
Returns two solution matrices, the low rank \code{L} matrix and the sparse \code{S} matrix.
}
\description{
\code{root_pcp_noncvx_na} implements \code{rootPCP} with NO non-negativity constraint on the \code{L} solution matrix and replaces the nuclear norm with a projection onto a smaller rank. \cr \cr
It solved the following ADMM splitting problem: \cr \cr
min(L,S) \cr
||L||_* + lambda * ||S||_1 + mu * || P_(obs)[L+S-D] ||_F \cr \cr
This is first transformed to the problem: \cr \cr
min(L1,L2,S1,S2,Z) \cr
||L1||_* + lambda * ||S1||_1 + mu * ||Z||_F \cr \cr
s.t. L1 = L2; S1 = S2; Z = P_obs[ D - L2 - S2]. \cr \cr
The algorithm conducts ADMM splitting as (L1,S1,Z),(L2,S2). \cr \cr
This version allows for missing values. \cr
Use NA for missing entries in D.
}
