% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sim_data}
\alias{sim_data}
\title{Simulate data}
\usage{
sim_data(
  n = 100,
  p = 10,
  r = 3,
  sparse_nonzero_idxs = NULL,
  sigma = 0.05,
  seed = 42
)
}
\arguments{
\item{n, p}{(Optional) A pair of integers specifying the simulated dataset's
number of \code{n} observations (rows) and \code{p} variables (columns). By default,
\code{n = 100}, and \code{p = 10}.}

\item{r}{(Optional) An integer specifying the rank of the simulated dataset's
low-rank component. Intuitively, the number of latent patterns governing
the simulated dataset. By default, \code{r} = 3.}

\item{sparse_nonzero_idxs}{(Optional) An integer vector specifying the
indices of the non-zero elements in the sparse component. By default,
\code{sparse_nonzero_idxs = NULL}, in which case it is defined to be the
vector \code{c(1, n * p)} (placing sparse noise in the very first and last
entries of the simulated dataset).}

\item{sigma}{(Optional) A double specifying the standard deviation of the
dense (Gaussian) noise component \code{Z}. By default, \code{sigma = 0.05}.}

\item{seed}{(Optional) An integer specifying the seed for random number
generation. By default, \code{seed = 42}.}
}
\value{
A list containing:
\itemize{
\item \code{D}: The observed data matrix, where \code{D = L + S + Z}.
\item \code{L}: The ground truth rank-\code{r} low-rank matrix.
\item \code{S}: The ground truth sparse matrix.
\item \code{S}: The ground truth dense (Gaussian) noise matrix.
}
}
\description{
\code{sim_data()} generates a simulated dataset \code{D = L + S + Z} for
experimentation with Principal Component Pursuit (PCP) algorithms.
}
\examples{
# rank 3 example
data <- sim_data()
matrix_rank(data$D)
matrix_rank(data$L)
# rank 7 example
data <- sim_data(n = 1000, p = 25, r = 7)
matrix_rank(data$D)
matrix_rank(data$L)
}
\seealso{
\code{\link[=corrupt_mat_randomly]{corrupt_mat_randomly()}}
}
