% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search_cv.R
\name{sparsity}
\alias{sparsity}
\title{Calculate the sparsity of a given matrix.}
\usage{
sparsity(S, c = 1e-05)
}
\arguments{
\item{S}{The input matrix whose sparsity is in question.}

\item{c}{(Optional) a numeric threshold / tolerance used to determine if an entry in \code{S}
is effectively 0. If the absolute value of an entry is below \code{c} then it is effectively 0.
By default, \code{c = 1e-05}.}
}
\value{
The sparsity of \code{S}.
}
\description{
\code{sparsity} computes the sparsity of a given matrix, \code{S}, using a given tolerance, \code{c}.
}
\examples{
sparsity(matrix(0:8, 3, 3))
}
