% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{loss_lod}
\alias{loss_lod}
\title{Loss for stablePCP-LOD}
\usage{
loss_lod(X, D, LOD)
}
\arguments{
\item{X}{The predicted value of \code{L + S} at the current iteration.}

\item{D}{The original dataset.}

\item{LOD}{The LOD. May be a scalar, vector (\code{length(LOD) = ncol(D)}), or matrix (\code{dim(LOD) == dim(D)}).}
}
\value{
Scalar value used to calculate loss in objective function.
}
\description{
\code{loss_lod} includes the LOD-specific penalty terms to compute the loss on the squared error term \code{L+S-D}.
}
