% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox_operators.R
\name{loss_lod}
\alias{loss_lod}
\title{Loss function for PCP's limit of detection (LOD) penalty.}
\usage{
loss_lod(D, X, LOD)
}
\arguments{
\item{D}{The original data matrix.}

\item{X}{The predicted value of \code{L + S} at the current iteration.}

\item{LOD}{The LOD as a matrix (\code{dim(LOD) == dim(D)}).}
}
\value{
Scalar value used to calculate loss in objective function.
}
\description{
\code{loss_lod} includes the LOD-specific penalty terms to compute the loss on the squared error term \code{L + S - D}.
}
\seealso{
\code{\link{stable_pcp}}
}
\keyword{internal}
