% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcpr-package.R
\docType{package}
\name{pcpr-package}
\alias{pcpr}
\alias{pcpr-package}
\title{pcpr: Principal Component Pursuit for Environmental Epidemiology}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implementation of the pattern recognition technique Principal Component Pursuit tailored to environmental health data, as described in Gibson et al. (2022) \doi{10.1289/EHP10479}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://columbia-prime.github.io/pcpr/}
  \item \url{https://github.com/Columbia-PRIME/pcpr}
  \item Report bugs at \url{https://github.com/Columbia-PRIME/pcpr/issues}
}

}
\author{
\strong{Maintainer}: Lawrence G. Chillrud \email{lawrencechillrud@gmail.com} (\href{https://orcid.org/0000-0003-0727-0161}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Jaime Benavides (\href{https://orcid.org/0000-0002-1851-5155}{ORCID})
  \item Elizabeth A. Gibson (\href{https://orcid.org/0000-0001-5119-5133}{ORCID})
  \item Junhui Zhang (\href{https://orcid.org/0009-0008-5922-1058}{ORCID})
  \item Jingkai Yan (\href{https://orcid.org/0000-0002-2094-2092}{ORCID})
  \item John N. Wright
  \item Jeff Goldsmith
  \item Marianthi-Anna Kioumourtzoglou (\href{https://orcid.org/0000-0001-5710-4992}{ORCID})
}

Other contributors:
\itemize{
  \item Columbia University (00hj8s172) [funder]
}

}
\keyword{internal}
