% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox_operators.R
\name{prox_frobenius}
\alias{prox_frobenius}
\title{Prox Frobenius}
\usage{
prox_frobenius(Z, c)
}
\arguments{
\item{Z}{The input error/noise matrix, \code{Z = D - L - S}.}

\item{c}{The amount by which the prox Frobenius method penalizes \code{Z}.}
}
\value{
The thresholded error/noise matrix.
}
\description{
\code{prox_frobenius} implements the proximal gradient method for the Frobenius norm.
This thresholding minimizes the square root of the sum of squared error.
The proximal gradient method is used to solve non-differentiable convex optimization problems.
This is only used in \code{root_pcp}, since the squared Frobenius error in \code{stable_pcp} is differentiable.
}
\seealso{
\code{\link{prox_l1}}, \code{\link{prox_nuclear}}
}
\keyword{internal}
