% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prox_l1}
\alias{prox_l1}
\title{Prox L1}
\usage{
prox_l1(Y, c)
}
\arguments{
\item{Y}{The \code{S} matrix.}

\item{c}{The amount by which the prox L1 method penalizes \code{Y}.}
}
\value{
The thresholded \code{S} matrix.
}
\description{
\code{prox_l1} implements the proximal gradient method for the L1 norm.
This soft thresholding encourages the \code{S} matrix to be sparse.
The proximal gradient method is used to solve non-differentiable convex optimization problems.
This is used in \code{stablePCP} and \code{rootPCP}.
}
