% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prox_nuclear}
\alias{prox_nuclear}
\title{Prox nuclear}
\usage{
prox_nuclear(Y, c)
}
\arguments{
\item{Y}{The \code{L} matrix.}

\item{c}{The amount by which the prox nuclear method penalizes \code{Y}.}
}
\value{
\code{prox_nuclear} returns two objects.
\code{X} is the thresholded \code{L} matrix.
\code{nuclearX} is the sum of the absolute values of the thresholded singular values.
This goes into the objective function.
}
\description{
\code{prox_nuclear} implements the proximal gradient method for the nuclear norm.
The nuclear norm is equivalent to the L2 norm of the singular values of the matrix.
This singular value thresholding encourages the \code{L} matrix to be low rank.
The proximal gradient method is used to solve non-differentiable convex optimization problems.
This is used in \code{stablePCP} and \code{rootPCP}.
}
