% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRMC.R
\name{RRMC}
\alias{RRMC}
\title{LGC: Objective function to report to user?? Non-negativity constraint? Needs code check!}
\usage{
RRMC(D, r, eta, LOD = -Inf)
}
\description{
LGC: Objective function to report to user?? Non-negativity constraint? Needs code check!
}
