% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_pcp.R
\name{stable_pcp}
\alias{stable_pcp}
\title{LGC: Need to add missing value functionality !! Needs a code check!}
\usage{
stable_pcp(
  D,
  lambda,
  mu,
  LOD = -Inf,
  non_negative = T,
  max_iter = 10000,
  verbose = F
)
}
\description{
LGC: Need to add missing value functionality !! Needs a code check!
}
\keyword{internal}
